VPD-1027 Fix the intermittent failures
The problem can occur when Mdc.preservingMdc() is passed a block that evaluates to a successful Future. The map() then runs on a thread from the ExecutionContext (which loses context) but the transformation explicitly preserves the MDC using putMdc(), which correctly preserves MDC to this point in the function. However, the recoverWith() then runs (despite doing nothing for a successful future) on a thread scheduler's choice; often this is the same thread, in which case the MDC is preserved accidentally, only if a different thread is used is the MDC not preserved.

A corresponding problem can occur when the block evaluates to a failed Future. The map() is run but does nothing (because the Future failed) and, depending on the thread chosen to do this, may or may not lose the MDC. The recoverWith() in this case always runs and explicitly preserves the MCD, but only if the MDC has made it past the map() transformation.

The scheduler very often uses the same thread for the map() and recoverWith() transformations, in which case we don't see a loss of MDC. However, once in a while the scheduler can use different threads, which explains why test failures are rarely seen.

Even when a different thread is chosen by the scheduler, a test may not fail if that thread has already been used in an earlier test. Mdc.preservingMdc() adds test data a test thread's MDC, but does not remove it afterwards. This leaves threads in the ExecutionContext that already have their MDC containing the original "k" -> "v" test data. To prevent this happening and expose the problem more often, these test values have been replaced by unique values in each test (UUIDs); to prevent incorrectly passing tests and show the failure more often. It is arguable that a further change Mdc.preservingMdc() to restore the original MDC of a thread after use ought to be implemented.

The fix in this commit uses a single transform() that applies to both a successful and unsuccessful Future. Each branch of this operation explicitly preserves the MDC and so there is no route whereby it might be lost. The soak tests that clearly showed the original issue now pass 100% of the tme.

In production systems, failure to preserve MDC won't occur because the thread pool being used by Mdc.preservingMdc() would have ensured that the transformations where MDC could be lost, preserve MDC.