VPD-1027 Fix intermittent test in the build pipeline
The test "Request.Mdc.add - should not lead to OOM" is trying to show that request data in MDC can be Garbage Collected when no longer referenced. It generates enough request data to force a GC and expects there to be fewer items in MDC than were entered. The previous number of requests, 20k, is sufficient to demonstrate the issue in a local dev environment where the heap is 1GB, but the build job is run with a JVM with 8GB, and so we need to size this appropriately for this to fail.